Feintuning

+++ Getraenk.cpp
@@ -1,7 +1,7 @@
 #include "Getraenk.hpp"
 
 
-Getraenk::Getraenk(float prozent, int volumen){
+Getraenk::Getraenk(float prozent, float volumen){
     AlkoholMasse =  (volumen * 0.8 * prozent)/100;
     volumenprozent = prozent;
 }
